리팩터링(상품) - queryHandler 분리
