+++ src/main/java/kitchenpos/products/tobe/application/ProductQueryHandler.java
@@ -1,0 +1,21 @@
+package kitchenpos.products.tobe.application;
+
+import java.util.List;
+import kitchenpos.products.tobe.domain.entity.Product;
+import kitchenpos.products.tobe.domain.repository.ProductRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+public class ProductQueryHandler {
+    private final ProductRepository productRepository;
+
+    public ProductQueryHandler(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Product> findAll() {
+        return productRepository.findAll();
+    }
+}
